#2 db

+++ src/main/java/com/springapp/rest/repository/StaticEmployeeDAO.java
@@ -9,17 +9,17 @@ import com.springapp.rest.model.Employee;
 
 public class StaticEmployeeDAO implements EmployeeDAO {
 	
-	static Map<Integer, Employee> empMap;
+	static Map<Long, Employee> empMap;
 	static{
-		empMap = new HashMap<Integer, Employee>();
-		Employee empOne = new Employee(1,"A","K","ak@gmail.com");
-	    Employee empTwo = new Employee(2,"B","K","bk@yahoo.com");
-	    Employee empThree = new Employee(3,"C","K","ck@gmail.com");
+		empMap = new HashMap<Long, Employee>();
+		Employee empOne = new Employee(1l,"A","K","ak@gmail.com");
+	    Employee empTwo = new Employee(2l,"B","K","bk@yahoo.com");
+	    Employee empThree = new Employee(3l,"C","K","ck@gmail.com");
 	     
 	     
-	    empMap.put(1,empOne);
-	    empMap.put(2,empTwo);
-	    empMap.put(3,(empThree));
+	    empMap.put(1l,empOne);
+	    empMap.put(2l,empTwo);
+	    empMap.put(3l,(empThree));
     }
 	
 	public StaticEmployeeDAO(){
@@ -28,7 +28,7 @@ public class StaticEmployeeDAO implements EmployeeDAO {
 	
 	
 	
-	public Employee getEmployee(Integer id) {
+	public Employee getEmployee(Long id) {
 		Employee emp = empMap.get(id);
 		return emp;
 	}
@@ -43,10 +43,11 @@ public class StaticEmployeeDAO implements EmployeeDAO {
 
 	}
 
-	public boolean deleteEmployee(Integer id) {
+	public boolean deleteEmployee(Long id) {
 		if(empMap.remove(id)!=null)
 			return true;
 		else return
 				false;
 	}
+
 }
